Moved to react-router v4
