+++ demo/package.json
@@ -38,6 +38,7 @@
     "classnames": "^1.2.0",
     "css-loader": "^0.23.1",
     "extract-text-webpack-plugin": "^1.0.1",
+    "history": "^4.3.0",
     "http-server": "^0.6.1",
     "karma": "~0.12.23",
     "karma-browserify": "^4.0.0",
@@ -50,6 +51,7 @@
     "postcss-loader": "^0.10.1",
     "process": "^0.11.2",
     "protractor": "~1.0.0",
+    "react-router": "^4.0.0-2",
     "sass-loader": "^3.2.0",
     "shelljs": "^0.2.6",
     "simple-assign": "^0.1.0",
@@ -91,7 +93,6 @@
     "react-addons-css-transition-group": "^15.0.1",
     "react-addons-transition-group": "^15.0.1",
     "react-dom": "^15.0.1",
-    "react-router": "^2.1.1",
     "ua-parser-js": "^0.7.9"
   }
 }
